# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api-server",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/admin/mailing-lists": {
      "get": {
        "tags": [
          "Mailing Lists"
        ],
        "description": "Get all mailing lists",
        "operationId": "routes_mailing_lists_list_mailing_lists",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_for_Array_of_MailingList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/mailing-lists/{slug}": {
      "get": {
        "tags": [
          "Mailing Lists"
        ],
        "description": "Get a specific mailing list by slug",
        "operationId": "routes_mailing_lists_get_mailing_list",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailingList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/mailing-lists/{slug}/repositories": {
      "get": {
        "tags": [
          "Mailing Lists"
        ],
        "description": "Get a mailing list with its repositories",
        "operationId": "routes_mailing_lists_get_mailing_list_with_repos",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailingListWithRepos"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/mailing-lists/{slug}/toggle": {
      "patch": {
        "tags": [
          "Mailing Lists"
        ],
        "description": "Toggle a mailing list enabled/disabled status",
        "operationId": "routes_mailing_lists_toggle_mailing_list",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/mailing-lists/seed": {
      "post": {
        "tags": [
          "Mailing Lists"
        ],
        "description": "Seed all mailing lists from lore.kernel.org manifest This endpoint is idempotent - safe to run multiple times Dynamically fetches and parses the grokmirror manifest",
        "operationId": "routes_mailing_lists_seed_mailing_lists",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DataResponse_for_Array_of_MailingList": {
        "description": "Simple data wrapper for non-paginated list endpoints",
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailingList"
            }
          }
        }
      },
      "MailingList": {
        "type": "object",
        "required": [
          "enabled",
          "id",
          "name",
          "slug",
          "sync_priority"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "sync_priority": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "MailingListWithRepos": {
        "type": "object",
        "required": [
          "enabled",
          "id",
          "name",
          "repos",
          "slug",
          "sync_priority"
        ],
        "properties": {
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailingListRepository"
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "sync_priority": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "MailingListRepository": {
        "type": "object",
        "required": [
          "id",
          "mailing_list_id",
          "repo_order",
          "repo_url"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mailing_list_id": {
            "type": "integer",
            "format": "int32"
          },
          "repo_url": {
            "type": "string"
          },
          "repo_order": {
            "type": "integer",
            "format": "int32"
          },
          "last_indexed_commit": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ToggleResponse": {
        "type": "object",
        "required": [
          "enabled",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ToggleRequest": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "SeedResponse": {
        "type": "object",
        "required": [
          "mailingListsCreated",
          "message",
          "partitionsCreated",
          "repositoriesCreated"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "mailingListsCreated": {
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "repositoriesCreated": {
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "partitionsCreated": {
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          }
        }
      },
...
    }
  }
}